api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.3",
    "firebase-admin>=7.1.0",
    "python-dotenv==1.2.1",
    "uvicorn>=0.38.0",
]
